Moving from one route to explicit routes in pubhub
